Initial Commit of StringConcatDelim UDF
Initial Commit of StringConcatDelim UDF, a version of StringConcat UDF,
which accepts delimiter when intitialized, and it also keeps the place
holder if null values are present. i.e. if any one of the relation
passed as argument to this UDF is null, the output is not null.It will
just ignore the NULL value and still concat rest of them.